Make Helper available starting from Go 1.7
