+++ is_helper_test.go
@@ -1,3 +1,5 @@
+// +build go1.7
+
 package is
 
 func helper(is *I, helper bool) {
