+++ is_test.go
@@ -299,38 +299,3 @@ func TestFormatStringEscape(t *testing.T) {
 		t.Errorf("string was not escaped correctly: %s", actual)
 	}
 }
-
-func TestHelper(t *testing.T) {
-	tests := []struct {
-		name             string
-		helper           bool
-		expectedFilename string
-	}{
-		{
-			name:             "without helper",
-			helper:           false,
-			expectedFilename: "is_helper_test.go",
-		},
-		{
-			name:             "with helper",
-			helper:           true,
-			expectedFilename: "is_test.go",
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			tt := &mockT{}
-			is := NewRelaxed(tt)
-
-			var buf bytes.Buffer
-			is.out = &buf
-			helper(is, tc.helper)
-			actual := buf.String()
-			t.Log(actual)
-			if !strings.Contains(actual, tc.expectedFilename) {
-				t.Errorf("string does not contain correct filename: %s", actual)
-			}
-		})
-	}
-}
